This a BookStore application. The project is split in library(dll) and Application
This is an example of how to handle with text and XML files. 
In order to increase, security user's password ware not saved in files but its hash instead.
This also an example of how to handle lists, listviews, comboxes, DateTimePickers and other controls
The project is structured in 3 levels, GUI, Bussiness (BLL) and DataAccess (DAL), what makes easier to change DAL layer to use Database instead of files.